# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog Post API",
        "description": "This is an API for a Blog site",
        "contact": {
            "name": "Jide Timson",
            "email": "timson.babajide@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/post": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "List all posts",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "A list of posts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PostData"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "New blog post",
                "operationId": "store",
                "requestBody": {
                    "description": "Post object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "A list of posts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostData"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/post/{id}": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Get blog post by ID",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the post you want to get",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Single Post",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostData"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Post"
                ],
                "summary": "Update blog post",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the post you want to get",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Post object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated post",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostData"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Post"
                ],
                "summary": "Delete blog post",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the post you want to get",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Post",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "integer"
                                },
                                "example": 1
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "PostData": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "title": {
                        "type": "string",
                        "example": "First Post"
                    },
                    "body": {
                        "type": "string",
                        "example": "This is my first post"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2022-08-22T21:06:02.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2022-08-22T21:06:02.000000Z"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PostData"
                }
            },
            "PostRequest": {
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "First Post"
                    },
                    "body": {
                        "type": "string",
                        "example": "This is my first post"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PostData"
                }
            }
        }
    }
}